hook: useInitApp
