+++ src/client/pages/User/User.tsx
@@ -4,6 +4,7 @@ import { useParams } from 'react-router-dom'
 import { Users } from '@meta/user'
 
 import { useAppDispatch } from '@client/store'
+import { useInitApp } from '@client/store/assessment/hooks/useInitApp'
 import { useUserToEdit } from '@client/store/ui//userManagement/hooks'
 import { UserManagementActions } from '@client/store/ui/userManagement'
 import { useUser } from '@client/store/user'
@@ -21,6 +22,8 @@ const User: React.FC = () => {
   const isAdministrator = Users.isAdministrator(user)
 
   const canEditUser = Users.getRolesAllowedToEdit({ user, countryIso }).length > 0
+
+  useInitApp()
 
   useEffect(() => {
     dispatch(UserManagementActions.getUserToEdit({ id: Number(userId), countryIso }))
